Making reading of time series more efficient

+++ projDir/qc/dataProcessing/timeSeries/calcMomentsTime.m
@@ -8,33 +8,33 @@ R2v=mean(cIQ.v(:,1:end-2).*conj(cIQ.v(:,3:end)),2);
 R3v=mean(cIQ.v(:,1:end-3).*conj(cIQ.v(:,4:end)),2);
 R4v=mean(cIQ.v(:,1:end-4).*conj(cIQ.v(:,5:end)),2);
 
-momentsTime.powerV(:,ii)=10*log10(R0v)-data.rx_gain_v;
-momentsTime.velRaw(:,ii)=data.lambda/(4*pi*mode(data.prt))*angle(R1v);
+momentsTime.powerV(:,ii)=single(10*log10(R0v)-data.rx_gain_v);
+momentsTime.velRaw(:,ii)=single(data.lambda/(4*pi*mode(data.prt))*angle(R1v));
 widthRaw=data.lambda/(2*pi.*mode(data.prt)*6^.5)*abs(log(abs(R1v./R2v))).^0.5;
-momentsTime.skew(:,ii)=abs(log(abs(R3v./(R2v.^3))));
-momentsTime.kurt(:,ii)=abs(log(abs(R4v./(R2v.^2))));
-momentsTime.ncp(:,ii)=abs(R1v)./R0v;
+momentsTime.skew(:,ii)=single(abs(log(abs(R3v./(R2v.^3)))));
+momentsTime.kurt(:,ii)=single(abs(log(abs(R4v./(R2v.^2)))));
+momentsTime.ncp(:,ii)=single(abs(R1v)./R0v);
 
 % Correct velocity for aircraft motion
 xCorr=sind(momentsTime.azimuth_vc(ii)).*cosd(momentsTime.elevation(ii)).*momentsTime.eastward_velocity(ii);
 yCorr=cosd(momentsTime.azimuth_vc(ii)).*cosd(momentsTime.elevation(ii)).*momentsTime.northward_velocity(ii);
 zCorr=sind(momentsTime.elevation(ii)).*momentsTime.vertical_velocity(ii);
-momentsTime.vel(:,ii)=momentsTime.velRaw(:,ii)+xCorr+yCorr+zCorr;
+momentsTime.vel(:,ii)=momentsTime.velRaw(:,ii)+single(xCorr+yCorr+zCorr);
 
 % Correct width for aircraft motion
 velAircraft=sqrt(momentsTime.eastward_velocity(ii).^2+momentsTime.northward_velocity(ii).^2);
 deltaC=0.3.*velAircraft.*sin(deg2rad(momentsTime.elevation(ii))).*deg2rad(data.beamwidth_v);
-momentsTime.width(:,ii)=real(sqrt(widthRaw.^2-deltaC.^2));
+momentsTime.width(:,ii)=single(real(sqrt(widthRaw.^2-deltaC.^2)));
 
 % SNRV
 noiseLinV=10.^(data.noise_v./10);
 snrLinV=(R0v-noiseLinV)./noiseLinV;
 snrLinV(snrLinV<0)=nan;
-momentsTime.snr(:,ii)=10*log10(snrLinV);
+momentsTime.snr(:,ii)=single(10*log10(snrLinV));
 
 % DBZV
 data.range(data.range<0)=nan;
-momentsTime.dbz(:,ii)=momentsTime.snr(:,ii)+20*log10(data.range./1000)+data.dbz1km_v;
+momentsTime.dbz(:,ii)=single(momentsTime.snr(:,ii)+20*log10(data.range./1000)+data.dbz1km_v);
 
 %% H
 if isfield(cIQ,'h')
@@ -42,7 +42,7 @@ if isfield(cIQ,'h')
     R0h=mean(real(cIQ.h).^2+imag(cIQ.h).^2,2);
 
     % Power H
-    momentsTime.powerH(:,ii)=10*log10(R0h)-data.rx_gain_h;
+    momentsTime.powerH(:,ii)=single(10*log10(R0h)-data.rx_gain_h);
 
     % SNRH
     noiseLinH=10.^(data.noise_h./10);
@@ -54,7 +54,7 @@ if isfield(cIQ,'h')
     dbzH=snrH+20*log10(data.range./1000)+data.dbz1km_h;
 
     % LDR
-    momentsTime.ldr(:,ii)=dbzH-momentsTime.dbz(:,ii);
+    momentsTime.ldr(:,ii)=single(dbzH-momentsTime.dbz(:,ii));
 end
 end
